Split out suggestion generation into parts
